Form changes

+++ app/Common.php
@@ -252,9 +252,11 @@ class Common
         if(!$files) rmdir($path);
         else{
             foreach ($files as $file) {
+            	if($file == '.' || $file == '..') continue;
                 if(is_dir($path . $file)) self::delete_folder($path . $file);
                 rmdir($path . $file);                
             }
+            rmdir($path);
         }
     }
